rename icds backend app -> icds_nic
