+++ corehq/apps/smsbillables/management/commands/bootstrap_icds_gateway.py
@@ -4,7 +4,7 @@ from corehq.apps.accounting.models import Currency
 from corehq.apps.sms.models import INCOMING, OUTGOING
 from corehq.apps.smsbillables.models import SmsGatewayFee, SmsGatewayFeeCriteria
 from corehq.apps.smsbillables.utils import log_smsbillables_info
-from corehq.messaging.smsbackends.icds.models import SQLICDSBackend
+from corehq.messaging.smsbackends.icds_nic.models import SQLICDSBackend
 
 
 def bootstrap_icds_gateway(apps=None):
